Reviewer 2
I got the project to run successfully, although it prompted me that there was an error on line 115, 1. For some reason it expected a comma in that location, I really don't understand why though. When I placed a comma there it ran as it was expected. However, while using my android emulator on genymotion the screen resolution didn't adjust to my phone size. Therefore, on the landing page the last letter "D" was cut off from the name.  Also, on the child page "Bio" they "Y" at the end of biography was cut off as well.  While reading through the code I couldn't find any other errors other than where it told me to put a comma. The project ran as intended on the iPad emulator though, and the screen resolution was adjusted accordingly.